requirements.txt
python-dotenv
elevenlabs
openai